wasd plugin: rename to keyremapping plugin
