+++ runelite-client/src/main/java/net/runelite/client/plugins/keyremapping/KeyRemappingListener.java
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.wasdcamera;
+package net.runelite.client.plugins.keyremapping;
 
 import com.google.common.base.Strings;
 import java.awt.event.KeyEvent;
@@ -37,13 +37,13 @@ import net.runelite.client.callback.ClientThread;
 import net.runelite.client.input.KeyListener;
 import net.runelite.client.input.MouseAdapter;
 
-class WASDCameraListener extends MouseAdapter implements KeyListener
+class KeyRemappingListener extends MouseAdapter implements KeyListener
 {
 	@Inject
-	private WASDCameraPlugin plugin;
+	private KeyRemappingPlugin plugin;
 
 	@Inject
-	private WASDCameraConfig config;
+	private KeyRemappingConfig config;
 
 	@Inject
 	private Client client;
